phaul: Prepare a socket for memory migration
Currently on each iteration page server opens a
listening socket and criu pre-dump/dump connects
to it.

This is wrong, criu should accept an FD to xfer
memory through from p.haul. To make this happen
p.haul must be able to establish a side channel
and this patch does so.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>