criu: Use swrk instead of remote service
The plan is to push page server socket to criu from
p.haul. This can be done by passing one with scm_rights,
but python2.2 cannot do it easily. Thus, for now, we
end up spawning the swrk, it will inherit all the fds
and we'll just "tell" one where the socket is.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>