criu: Manually handle fds passed (and closed) to CRIU
With the self-made RPC we can explicitly mark the
unwanted descriptors with cloexec flag and control
which FDs go to CRIU swrk.

This saves us from having control socket in stdin
and memory socket in stdout. And keeps stdion open
so we can see logs from CRIU on the screen.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>